api/requirements.txt
aio-pika==9.4.1
celery==5.4.0
fastapi[all]==0.110.2
minio==7.2.6
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
sqlmodel==0.0.16